requirements.txt
selenium==3.3.1
pytest==3.1.1
Faker==0.8.3
allure-pytest==2.2.2b2
simple-settings==0.16.0